FIX bad XREPEAT bug, add INC & DEC
XREPEAT had a bad bug where a trialing / terminating separator was
output for counts (?)(0), (?)(0)(0), (?)(0)(0)(0).
The fix is to implement XREPEAT in terms of IREPEAT.
This required a way to decrement the HEXS index, as well as testing
For all (0) HEXS and decrementing it to nothing - empty.
Both XREPEAT and IREPEAT now accept empty for the HEXS count.